% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AdverseEvents.R
\name{Adverse Events Domain}
\alias{Adverse Events Domain}
\title{Adverse Events Domain}
\description{
The Adverse Events data of an ImmPort study is reformated to the CDISC SDTM Adverse 
Events (AE) domain model, and is a list of 2 data frames containing 1) Adverse Events data \code{\link{AE}} 
and 2) any supplemental Adverse Events data \code{\link{SUPP}}
}

