% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AdverseEvents.R
\name{AE}
\alias{AE}
\title{Adverse Events Domain Variables}
\description{
{
  \tabular{ll}{
    \strong{Variable Name  } \tab \strong{Variable Label} \cr
    STUDYID \tab Study Identifier \cr
    DOMAIN  \tab Domain Abbreviation \cr
    USUBJID \tab Unique Subject Identifier \cr
    AESEQ \tab Sequence Number \cr
    AESPID \tab Sponsor-Defined Identifier \cr
    AETERM \tab Reported Term for the Adverse Event \cr
    AEMODIFY \tab Modified Reported Term \cr
    AEBODYSYS \tab Body System or Organ Class \cr
    AELOC \tab Location of Event \cr
    AESEV \tab Severity/Intensity \cr
    AESER \tab Serious Event \cr
    AEACN \tab Action Taken with Study Treatment \cr
    AEACNOTH \tab Other Action Taken \cr
    AEREL \tab Causality \cr
    AERELNST \tab Relationship to Non-Study Treatment \cr
    AEOUT \tab Outcome of Adverse Event \cr
    AESTDY \tab Study Day of Start of Adverse Event \cr
    AEENDY \tab Study Day of End of Adverse Event
  }
}
}

