% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AssociatedPersonsMedicalHistory.R
\name{APMH}
\alias{APMH}
\title{Associated Persons Medical History Domain Variables}
\description{
{
  \tabular{ll}{
    \strong{Variable Name  } \tab \strong{Variable Label} \cr
    
    STUDYID \tab Study Identifier \cr
    DOMAIN  \tab Domain Abbreviation \cr
    APID \tab Associated Persons Identifier \cr
    MHSEQ \tab Sequence Number \cr
    RSUBJID \tab Related Subject \cr
    SREL \tab Subject Relationship \cr
    MHTERM \tab Reported Term for the Medical History \cr
    MHCAT \tab Category for Medical History \cr
    MHBODSYS \tab Body System or Organ Class \cr
    MHDTC \tab Date/Time of History Collection \cr
    MHDY \tab RStudy Day of History Collection
  }
}
}

