% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AssociatedPersonsMedicalHistory.R
\name{Associated Persons Medical History Domain}
\alias{Associated Persons Medical History Domain}
\title{Associated Persons Medical History Domain}
\description{
The Associated Persons Medical History data of an ImmPort study is reformated to the CDISC 
SDTM AAssociated Persons Medical History (APMH) domain model, and is a list of 2 data frames 
containing 1) Associated Persons Medical History data \code{\link{APMH}} and 2) any supplemental 
Associated Persons Medical History data \code{\link{SUPP}}
}

