% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ImmPortSqlite.R
\name{buildNewSqliteDb}
\alias{buildNewSqliteDb}
\title{buildNewSqliteDb}
\usage{
buildNewSqliteDb(data_dir, db_dir)
}
\arguments{
\item{data_dir}{File directory where the study TSV files are stored}

\item{db_dir}{File directory where the sqlite database will be stored}
}
\description{
The function \code{buildSqliteDb} builds a sqlite db of ImmPort study data. It takes in as input the
study files in the TSV (Tab) format.
}
\examples{
\dontrun{
  buildNewSqliteDb(data_dir, db_dir)
}
}

