% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CellularQuantification.R
\name{Cellular Quantification Domain}
\alias{Cellular Quantification Domain}
\title{Cellular Quantification Domain}
\description{
The ImmPort study data generated from assays of types: Flow and ELISPOT are grouped into the 
Cellular Quantification Domain. The data is reformated to a custom Cellular Quantification domain model in 
CDISC SDTM standards, and is a list of 2 data frames containing 1) Cellular Quantification data \code{\link{ZB}} and 2) any supplemental 
  Cellular Quantification data \code{\link{SUPP}}
}

