% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ConcomitantMedications.R
\name{CM}
\alias{CM}
\title{Concomitant Medications Domain Variables}
\description{
{
  \tabular{ll}{
    \strong{Variable Name  } \tab \strong{Variable Label} \cr
    STUDYID \tab Study Identifier \cr
    DOMAIN  \tab Domain Abbreviation \cr
    USUBJID \tab Unique Subject Identifier \cr
    CMSEQ \tab Sequence Number \cr
    CMTRT \tab Reported Name of Drug, Med, or Therapy \cr
    CMCAT \tab Category for Medication \cr
    CMDOSE \tab Dose per Administration \cr
    CMDOSTXT \tab Dose Description \cr
    CMDOSU \tab Dose Units \cr
    CMDOSFREQ \tab Dosing Frequency per Interval  \cr
    CMROUTE \tab Route of Administration \cr
    CMSTDTC \tab Start Date/Time of Medication \cr
    CMENDTC \tab End Date/Time of Medication \cr
    CMSTDY \tab Study Day of Start of Medication \cr
    CMENDY \tab Study Day of End of Medication
  }
}
}

