% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ConcomitantMedications.R
\name{Concomitant Medications Domain}
\alias{Concomitant Medications Domain}
\title{Concomitant Medications Domain}
\description{
The Concomitant Medications data of an ImmPort study is reformated to the CDISC SDTM 
Concomitant Medications (CM) domain model, and is a list of 2 data frames containing 
1) Concomitant Medications data \code{\link{CM}} and 2) any supplemental Concomitant 
Medications data \code{\link{SUPP}}
}

