% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Demographics.R
\name{Demographics Domain}
\alias{Demographics Domain}
\title{Demographics Domain}
\description{
The Demographics data of an ImmPort study is reformated to the CDISC SDTM Demographics (DM) 
domain model, and is a list of 2 data frames containing 1) Demographics data \code{\link{DM}} 
and 2) any supplemental Demographics data \code{\link{SUPP}}
}

