% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Demographics.R
\name{DM}
\alias{DM}
\title{Demographics Domain Variables}
\description{
{
  \tabular{ll}{
    \strong{Variable Name  } \tab \strong{Variable Label} \cr
    STUDYID \tab Study Identifier \cr
    DOMAIN  \tab Domain Abbreviation \cr
    USUBJID \tab Unique Subject Identifier \cr
    AGE \tab Age \cr
    AGEU \tab Age Units \cr
    SEX \tab Sex \cr
    RACE \tab Race (only for human data) \cr
    ETHNIC \tab Ethnicity (only for human data) \cr
    SPECIES \tab Species (only for non-human data) \cr
    STRAIN \tab Strain/Substrain (only for non-human data) \cr
    SBSTRAIN \tab Strain/Substrain Details (only for non-human data) \cr
    ARMCD \tab Planned Arm Code \cr
    ARM \tab Description of Planned Arm
  }
}
}

