% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ProtocolDeviations.R
\name{DV}
\alias{DV}
\title{Protocol Deviations Domain Variables}
\description{
{
  \tabular{ll}{
    \strong{Variable Name  } \tab \strong{Variable Label} \cr
    STUDYID \tab Study Identifier \cr
    DOMAIN  \tab Domain Abbreviation \cr
    USUBJID \tab Unique Subject Identifier \cr
    DVSEQ \tab Sequence Number \cr
    DVTERM \tab Protocol Deviation Term
  }
}
}

