% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Study.R
\docType{class}
\name{Events-class}
\alias{Events}
\alias{Events-class}
\title{Events class}
\description{
Events class
}
\section{Fields}{

\describe{
\item{\code{ae_l}}{Adverse Events data \code{\link{AE}} and supplemental Adverse Events data \code{\link{SUPP}}}

\item{\code{dv_l}}{Protocol Deviations data \code{\link{DV}} and supplemental Protocol Deviations data \code{\link{SUPPDV}}}

\item{\code{mh_l}}{Medical History data \code{\link{MH}} and supplemental Medical History data \code{\link{SUPPMH}}}

\item{\code{apmh_l}}{Associated Persons Medical History data \code{\link{APMH}} and supplemental Associated Persons Medical History data \code{\link{SUPP}}}
}}

