% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Exposure.R
\name{EX}
\alias{EX}
\title{Exposure Domain Variables}
\description{
{
  \tabular{ll}{
    \strong{Variable Name  } \tab \strong{Variable Label} \cr
    STUDYID \tab Study Identifier \cr
    DOMAIN  \tab Domain Abbreviation \cr
    USUBJID \tab Unique Subject Identifier \cr
    EXSEQ \tab Sequence Number \cr
    EXTRT \tab Name of Treatment \cr
    EXCAT \tab Category of Treatment \cr
    EXDOSE \tab Dose \cr
    EXDOSTXT \tab Dose Description \cr
    EXDOSU \tab Dose Units \cr
    EXDOSFRQ \tab Dosing Frequency per Interval \cr
    EXROUTE \tab Route of Administration \cr
    EXSTDTC \tab Start Date/Time of Treatment \cr
    EXENDTC \tab End Date/Time of Treatment \cr
    EXSTDY \tab Study Day of Start of Treatment \cr
    EXENDY \tab Study Day of End of Treatment
  }
}
}

