% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Exposure.R
\name{Exposure Domain}
\alias{Exposure Domain}
\title{Exposure Domain}
\description{
The Exposure data of an ImmPort study is reformated to the CDISC SDTM Exposure (EX) 
domain model, and is a list of 2 data frames containing 1) Exposure data \code{\link{EX}} 
and 2) any supplemental Exposure data \code{\link{SUPP}}
}

