% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FindingsAbout.R
\name{FA}
\alias{FA}
\title{Findings About Domain Variables}
\description{
{
  \tabular{ll}{
    \strong{Variable Name  } \tab \strong{Variable Label} \cr
    STUDYID \tab Study Identifier \cr
    DOMAIN  \tab Domain Abbreviation \cr
    USUBJID \tab Unique Subject Identifier \cr
    FASEQ \tab Sequence Number \cr
    FATEST \tab Findings About Test Name \cr
    FAOBJ \tab Object of the Observation \cr
    FACAT \tab Category for Findings About \cr
    FAORRES \tab Results or Findings in Original Units \cr
    FAORRESU \tab Original Units \cr
    FALOC \tab Location of the Finding About \cr
    VISITNUM \tab Visit Number \cr
    VISIT \tab Visit Name \cr
    FADY \tab Study Day of Collection
  }
}
}

