% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FindingsAbout.R
\name{Findings About Domain}
\alias{Findings About Domain}
\title{Findings About Domain}
\description{
The Findings About data of an ImmPort study is reformated to the CDISC SDTM Findings 
About (FA) domain model, and is a list of 2 data frames containing 1) Findings About data \code{\link{FA}} 
and 2) any supplemental Findings About data \code{\link{SUPPFA}}
}

