% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Study.R
\docType{class}
\name{Findings-class}
\alias{Findings}
\alias{Findings-class}
\title{Findings class}
\description{
Findings class
}
\section{Fields}{

\describe{
\item{\code{lb_l}}{Laboratory Test Results data \code{\link{LB}} and supplemental Laboratory Test Results data \code{\link{SUPPLB}}}

\item{\code{pe_l}}{Physical Examination data \code{\link{PE}} and supplemental Physical Examination data \code{\link{SUPPPE}}}

\item{\code{vs_l}}{Vital Signs data \code{\link{VS}} and supplemental Vital Signs data \code{\link{SUPPVS}}}

\item{\code{qs_l}}{Questionnaires data \code{\link{QS}} and supplemental Questionnaires data \code{\link{SUPP}}}

\item{\code{fa_l}}{Findings About data \code{\link{FA}} and supplemental Findings About data \code{\link{SUPPFA}}}

\item{\code{sr_l}}{Skin Response data \code{\link{SR}} and supplemental Skin Response data \code{\link{SUPP}}}

\item{\code{pf_l}}{Genetics Findings data \code{\link{PF}} and supplemental Genetics Findings data \code{\link{SUPPPF}}}

\item{\code{za_l}}{Protein Quantification data \code{\link{ZA}} and supplemental Protein Quantification data \code{\link{SUPPZA}}}

\item{\code{zb_l}}{Cellular Quantification data \code{\link{ZB}} and supplemental Cellular Quantification data \code{\link{SUPP}}}

\item{\code{zc_l}}{Nucleic Acid Quantification data \code{\link{ZC}} and supplemental Nucleic Acid Quantification data \code{\link{SUPP}}}

\item{\code{zd_l}}{Titer Assay Results data \code{\link{ZD}} and supplemental Titer Assay Results data \code{\link{SUPP}}}
}}
\examples{
library(DBI)
library(sqldf)
studies_dir <- system.file("extdata", "ImmPortStudies", package = "RImmPort")
db_dir <- file.path(studies_dir, "Db")
sqlite_conn <- dbConnect(SQLite(), dbname=file.path(db_dir, "ImmPort.sqlite"))
setImmPortDataSource(sqlite_conn)
sdy139 <- getStudy("SDY139")
zb_df <- sdy139$findings$zb_l$zb_df
}

