% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GeneticsFindings.R
\name{Genetics Findings Domain}
\alias{Genetics Findings Domain}
\title{Genetics Findings Domain}
\description{
The ImmPort study data generated from assays of types: HLA Typing and Array are grouped into the 
Genetics Findings Domain. The data is reformated to  a PharmacoGenomics and Genetics Findings (PF) model in 
CDISC SDTM standards, and is a list of 2 data frames containing 1) Genetics Findings data \code{\link{PF}} 
and 2) any supplemental Genetics Findings data \code{\link{SUPP}}
}

