% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Study.R
\name{getAssayDataOfStudies}
\alias{getAssayDataOfStudies}
\title{Get specific assay data of one or more studies from the ImmPort database}
\usage{
getAssayDataOfStudies(study_ids, assay_type)
}
\arguments{
\item{study_ids}{List of study indentifiers}

\item{assay_type}{Assay Type}
}
\value{
a list of 1) domain data of specicifc assay technology and 2) any supplemental domain data of the studies
}
\description{
Get specific assay data of one or more studies from the ImmPort database
}
\examples{
\dontrun{ 
    data_l <- getAssayDataOfStudies("SDY1", "ELISA")
    data_l <- getAssayDataOfStudies(c("SDY1", "SDY2"), "Flow")
  }
}
\author{
Ravi Shankar
}

