% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Study.R
\name{getDomainCode}
\alias{getDomainCode}
\title{Get code of a specific domain}
\usage{
getDomainCode(domain)
}
\arguments{
\item{domain}{Name of a specific domain}
}
\value{
A list of of all domain names and codes
}
\description{
The function \code{getListOfDomains} returns the code of a specific domain
}
\examples{
 domain <- "Demographics"
 code <- getDomainCode(domain)
}

