% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Study.R
\name{getListOfAssayTypes}
\alias{getListOfAssayTypes}
\title{Get a list of Assay Types}
\usage{
getListOfAssayTypes()
}
\value{
A list of assay types
}
\description{
The function \code{getListOfAssayTypes} returns a list of assay types that ImmPort studies have employed in their experimetal assays
}
\examples{
 at_l <- getListOfAssayTypes()
}

