% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Study.R
\name{getListOfDomains}
\alias{getListOfDomains}
\title{Get names of all domains and their codes}
\usage{
getListOfDomains()
}
\value{
A list of of all domain names and codes
}
\description{
The function \code{getListOfDomains} returns a list of all domain names and codes
}
\examples{
 domains_df <- getListOfDomains()
}

