% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Study.R
\name{getListOfStudies}
\alias{getListOfStudies}
\title{Get a list of all studies}
\usage{
getListOfStudies()
}
\value{
List of study indentifiers
}
\description{
Get a list of all studies
}
\examples{
\dontrun{ 
  study_ids <- getListOfStudies()
  }
}

