% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Study.R
\name{getStudiesWithSpecificDomainData}
\alias{getStudiesWithSpecificDomainData}
\title{Get a list of studies that have specific domain data}
\usage{
getStudiesWithSpecificDomainData(domain)
}
\arguments{
\item{domain}{Name of a specific domain}
}
\value{
List of study indentifiers
}
\description{
Get a list of studies that have specific domain data
}
\examples{
\dontrun{
   study_ids <- getStudiesWithSpecificDomainData("Demographics")
   study_ids <- getStudiesWithSpecificDomainData("Adverse Events")
}
}

