% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Study.R
\docType{class}
\name{Interventions-class}
\alias{Interventions}
\alias{Interventions-class}
\title{Interventions class}
\description{
Interventions class
}
\section{Fields}{

\describe{
\item{\code{cm_l}}{Concomitant Medications data \code{\link{CM}} and supplemental Concomitant Medications data \code{\link{SUPP}}}

\item{\code{ex_l}}{Exposure data \code{\link{EX}} and supplemental Exposure data \code{\link{SUPP}}}

\item{\code{su_l}}{Substance Use data \code{\link{SU}} and supplemental Substance Use data \code{\link{SUPP}}}
}}

