% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LaboratoryTestResults.R
\name{Laboratory Test Results Domain}
\alias{Laboratory Test Results Domain}
\title{Laboratory Test Results Domain}
\description{
The Laboratory Test Results data of an ImmPort study is reformated to the CDISC SDTM 
Laboratory Test Results (LB) domain model, and is a list of 2 data frames containing 1) Laboratory 
Test Results data \code{\link{LB}} and 2) any supplemental Laboratory Test Results data \code{\link{SUPPLB}}
}

