% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LaboratoryTestResults.R
\name{LB}
\alias{LB}
\title{Laboratory Test Results Domain Variables}
\description{
{
  \tabular{ll}{
    \strong{Variable Name  } \tab \strong{Variable Label} \cr
    STUDYID \tab Study Identifier \cr
    DOMAIN  \tab Domain Abbreviation \cr
    USUBJID \tab Unique Subject Identifier \cr
    LBSEQ \tab Sequence Number \cr
    LBTEST \tab Lab Test or Examination Name \cr
    LBCAT \tab Category for Lab Test \cr
    LBORRES \tab Result or Finding in Original Units \cr
    LBORRESU \tab Original Units \cr
    LBSPEC \tab Specimen Type \cr
    LBREFID \tab Specimen Identifier \cr
    VISITNUM \tab Visit Number \cr
    VISIT \tab Visit Name \cr
    LBELTM \tab Planned Elapsed Time from Time Point Ref \cr
    LBTPTREF \tab Time Point Reference
  }
}
}

