% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MedicalHistory.R
\name{Medical History Domain}
\alias{Medical History Domain}
\title{Medical History Domain}
\description{
The Medical History data of an ImmPort study is reformated to the CDISC SDTM Medical 
History (MH) domain model, and is a list of 2 data frames containing 1) Medical History 
data \code{\link{MH}} and 2) any supplemental Medical History data \code{\link{SUPPMH}}
}

