% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MedicalHistory.R
\name{MH}
\alias{MH}
\title{Medical History Domain Variables}
\description{
{
  \tabular{ll}{
    \strong{Variable Name  } \tab \strong{Variable Label} \cr
    
    STUDYID \tab Study Identifier \cr
    DOMAIN  \tab Domain Abbreviation \cr
    USUBJID \tab Unique Subject Identifier \cr
    MHSEQ \tab Sequence Number \cr
    MHTERM \tab Reported Term for the Medical History \cr
    MHCAT \tab Category for Medical History \cr
    MHBODSYS \tab Body System or Organ Class \cr
    MHDY \tab Study Day of History Collection
  }
}
}

