% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NucleicAcidQuantification.R
\name{Nucleic Acid Quantification Domain}
\alias{Nucleic Acid Quantification Domain}
\title{Nucleic Acid Quantification Domain}
\description{
The ImmPort study data generated from assays of types: PCR are grouped into the 
Nucleic Acid Quantification Domain. The data is reformated to a custom Nucleic Acid Quantification domain model in 
CDISC SDTM standards, and is a list of 2 data frames containing 1) Nucleic Acid Quantification data \code{\link{ZC}} and 2) any supplemental 
  Nucleic Acid Quantification data \code{\link{SUPP}}
}

