% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GeneticsFindings.R
\name{PF}
\alias{PF}
\title{Genetics Findings Domain Variables}
\description{
{
  \tabular{ll}{
    \strong{Variable Name  } \tab \strong{Variable Label} \cr
    STUDYID \tab Study Identifier \cr
    DOMAIN  \tab Domain Abbreviation \cr
    USUBJID \tab Unique Subject Identifier \cr
    PFSEQ \tab Sequence Number \cr
    PFGRPID \tab Group Identifier \cr
    PFTEST \tab  Test Name \cr
    PFCAT \tab Category for Test \cr
    PFMETHOD \tab Method of the Test \cr
    PFGENRI \tab Genetic Region of Interest \cr
    PFORRES \tab Result or Finding in Original Units \cr
    PFALLELC \tab Allele (Chromosome) Identifier \cr
    PFSPEC \tab Specimen Type \cr
    PFREFID \tab Reference ID (Specimen Identifier) \cr
    VISITNUM \tab Visit Number \cr
    VISIT \tab Visit Name \cr
    PFELTM \tab Planned Elapsed Time from Time Point Ref \cr
    PFTPTREF \tab Time Point Reference
  }
}
}

