% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PhysicalExamination.R
\name{Physical Examination Domain}
\alias{Physical Examination Domain}
\title{Physical Examination Domain}
\description{
The Physical Examination data of an ImmPort study is reformated to the CDISC SDTM 
Physical Examination (PE) domain model, and is a list of 2 data frames containing 1) Physical 
Examination data \code{\link{PE}} and 2) any supplemental Physical Examination data \code{\link{SUPPPE}}
}

