% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ProteinQuantification.R
\name{Protein Quantification Domain}
\alias{Protein Quantification Domain}
\title{Protein Quantification Domain}
\description{
The ImmPort study data generated from assays of types: ELISA and MBAA are grouped into the 
Cellular Quantification Domain. The data is reformated to a custom Protein Quantification domain model in 
CDISC SDTM standards, and is a list of 2 data frames containing 1) Protein Quantification data \code{\link{ZA}} 
and 2) any supplemental Protein Quantification data \code{\link{SUPP}}
}

