% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ProtocolDeviations.R
\name{Protocol Deviations Domain}
\alias{Protocol Deviations Domain}
\title{Protocol Deviations Domain}
\description{
The Protocol Deviations data of an ImmPort study is reformated to the CDISC SDTM 
Protocol Deviations (DV) domain model, and is a list of 2 data frames containing 1) Protocol 
Deviations data \code{\link{DV}} and 2) any supplemental Protocol Deviations data \code{\link{SUPPDV}}
}

