% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Questionnaires.R
\name{QS}
\alias{QS}
\title{Questionnaires Domain Variables}
\description{
{
  \tabular{ll}{
    \strong{Variable Name  } \tab \strong{Variable Label} \cr
    STUDYID \tab Study Identifier \cr
    DOMAIN  \tab Domain Abbreviation \cr
    USUBJID \tab Unique Subject Identifier \cr
    QSSEQ \tab Sequence Number \cr
    QSTEST \tab Questionnaires Test Name \cr
    QSCAT \tab Category for Questionnaires \cr
    QSORRES \tab Results or Findings in Original Units \cr
    QSORRESU \tab Original Units \cr
    VISITNUM \tab Visit Number \cr
    VISIT \tab Visit Name \cr
    QSDY \tab Study Day of Finding
  }
}
}

