% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Questionnaires.R
\name{Questionnaires Domain}
\alias{Questionnaires Domain}
\title{Questionnaires Domain}
\description{
The Questionnaires data of an ImmPort study is reformated to the CDISC SDTM 
Questionnaires (QS) domain model, and is a list of 2 data frames containing 1) Questionnaires 
data \code{\link{QS}} and 2) any supplemental Questionnaires data \code{\link{SUPP}}
}

