% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RImmPort.R
\docType{package}
\name{RImmPort}
\alias{RImmPort}
\alias{RImmPort-package}
\title{RImmPort: Enabling ready-for-analysis immunology research data}
\description{
The `RImmPort` package simplifies access to ImmPort data for analysis, 
as the name implies, in the R statistical environment. It provides a 
standards-based interface to the ImmPort study data that is in a proprietary format.
}

