% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Study.R
\name{serialzeStudyData}
\alias{serialzeStudyData}
\title{Serialize the Study Data}
\usage{
serialzeStudyData(study_ids, data_dir)
}
\arguments{
\item{study_ids}{List of study indentifiers}

\item{data_dir}{Path to a file folder where the .rds study files will be saved into}
}
\description{
Load specific studies from the database and save it in .rds format in 
a local file directory
}
\examples{
\dontrun{ 
  serialzeStudyData(study_ids, data_dir)
}

}

