% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SkinResponse.R
\name{Skin Response Domain}
\alias{Skin Response Domain}
\title{Skin Response Domain}
\description{
The Skin Response data of an ImmPort study is reformated to the CDISC SDTM Skin Response (SR) 
domain model, and is a list of 2 data frames containing 1) Skin Response data \code{\link{SR}} 
and 2) any supplemental Skin Response data \code{\link{SUPP}}
}

