% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Study.R
\docType{class}
\name{Study-class}
\alias{Study}
\alias{Study-class}
\title{Study class}
\description{
Study class
}
\section{Fields}{

\describe{
\item{\code{special_purpose}}{\code{\link{SpecialPurpose}}}

\item{\code{interventions}}{\code{\link{Interventions}}}

\item{\code{events}}{\code{\link{Events}}}

\item{\code{findings}}{\code{\link{Findings}}}

\item{\code{trial_design}}{\code{\link{TrialDesign}}}
}}

