% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SubstanceUse.R
\name{SU}
\alias{SU}
\title{Substance Use Domain Variables}
\description{
{
  \tabular{ll}{
    \strong{Variable Name  } \tab \strong{Variable Label} \cr
    STUDYID \tab Study Identifier \cr
    DOMAIN  \tab Domain Abbreviation \cr
    USUBJID \tab Unique Subject Identifier \cr
    SUSEQ \tab Sequence Number \cr
    SUTRT \tab Reported Name of Substance \cr
    SUCAT \tab Category of Substance Use \cr
    SUDOSE \tab Substance Use Consumption \cr
    SUDOSTXT \tab Substance Use Consumption Text \cr
    SUDOSU \tab Consumption Units \cr
    SUDOSFRQ \tab Use Frequency per Interval \cr
    SUROUTE \tab Route of Administration \cr
    SUSTDTC \tab Start Date/Time of Substance Use \cr
    SUENDTC \tab End Date/Time of Substance Use \cr
    SUSTDY \tab Study Day of Start of Substance Use \cr
    SUENDY \tab Study Day of End of Substance Use
  }
}
}

