% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SubjectVisits.R
\name{Subject Visits Domain}
\alias{Subject Visits Domain}
\title{Subject Visits Domain}
\description{
The Demographics data of an ImmPort study is reformated to the CDISC SDTM Subject Visits (SV) 
domain model, and is a list of 2 data frames containing 1) Subject Visits data \code{\link{SV}} 
and 2) any supplemental Subject Visits data \code{\link{SUPP}}
}

