% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SubstanceUse.R
\name{Substance Use Domain}
\alias{Substance Use Domain}
\title{Substance Use Domain}
\description{
The Substance Use data of an ImmPort study is reformated to the CDISC SDTM Substance Use (SU) 
domain model, and is a list of 2 data frames containing 1) Substance Use data \code{\link{SU}} 
and 2) any supplemental Substance Use data \code{\link{SUPP}}
}

