% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Demographics.R
\name{SUPP}
\alias{SUPP}
\title{Supplemental Variables}
\description{
{
  \tabular{ll}{
    \strong{Variable Name} \tab \strong{Variable Label} \cr
    STUDYID \tab Study Identifier \cr
    RDOMAIN  \tab Related Domain Abbreviation \cr
    USUBJID \tab Unique Subject Identifier \cr
    IDVAR \tab Identifying Variable \cr
    IDVARVAL \tab Identifying Variable Value \cr
    QNAM \tab Qualifier Variable Name \cr
    QLABEL \tab Qualifier Variable Label \cr
    QVAL \tab Data Value
  }
}
}

