% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ProtocolDeviations.R
\name{SUPPDV}
\alias{SUPPDV}
\title{Protocol Deviations Domain Supplemental Variables}
\description{
{
  \tabular{ll}{
    \strong{Variable Name} \tab \strong{Variable Label} \cr
    STUDYID \tab Study Identifier \cr
    RDOMAIN  \tab Related Domain Abbreviation \cr
    USUBJID \tab Unique Subject Identifier \cr
    IDVAR \tab Identifying Variable \cr
    IDVARVAL \tab Identifying Variable Value \cr
    QNAM \tab Qualifier Variable Name \cr
    QLABEL \tab Qualifier Variable Label \cr
    QVAL \tab Data Value
  }
}
}
\note{
The following table enumerates the values in QNAM and QLABEL variables {
  \tabular{ll}{
    \strong{QNAM} \tab \strong{QLABEL} \cr
    DVRELAE \tab Is Deviation Related to an Adverse Event? \cr
    DVREASON \tab Reason for Deviation \cr
    DVRESOL \tab Resulotion of Deviation \cr
    DVCONT \tab Did Subject continued in Study? \cr
    DVSTDY \tab Study Day of Start of Deviation \cr
    DVENDY \tab Study Day of End of Deviation
  }
}
}

