% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FindingsAbout.R
\name{SUPPFA}
\alias{SUPPFA}
\title{Findings About Domain Supplemental Variables}
\description{
{
  \tabular{ll}{
    \strong{Variable Name} \tab \strong{Variable Label} \cr
    STUDYID \tab Study Identifier \cr
    RDOMAIN  \tab Related Domain Abbreviation \cr
    USUBJID \tab Unique Subject Identifier \cr
    FASEQ \tab Sequence Number \cr
    IDVAR \tab Identifying Variable \cr
    IDVARVAL \tab Identifying Variable Value \cr
    QNAM \tab Qualifier Variable Name \cr
    QLABEL \tab Qualifier Variable Label \cr
    QVAL \tab Data Value
  }
}
}
\note{
The following table enumerates the values in QNAM and QLABEL variables {
  \tabular{ll}{
    \strong{QNAM} \tab \strong{QLABEL} \cr
    FATOD \tab Time of Day of Collection \cr
  }
}
}

