% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/GeneticsFindings.R
\name{SUPPPF}
\alias{SUPPPF}
\title{Genetics Findings Domain Supplemental Variables}
\description{
{
  \tabular{ll}{
    \strong{Variable Name} \tab \strong{Variable Label} \cr
    STUDYID \tab Study Identifier \cr
    RDOMAIN  \tab Related Domain Abbreviation \cr
    USUBJID \tab Unique Subject Identifier \cr
    IDVAR \tab Identifying Variable \cr
    IDVARVAL \tab Identifying Variable Value \cr
    QNAM \tab Qualifier Variable Name \cr
    QLABEL \tab Qualifier Variable Label \cr
    QVAL \tab Data Value
  }
}
}
\note{
The following table enumerates the values in QNAM and QLABEL variables {
  \tabular{ll}{
    \strong{QNAM} \tab \strong{QLABEL} \cr
    PFPOPAR \tab Geographic Area of the Population \cr
    PFSPECSB \tab Specimen Subtype
  }
}
}

