% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CellularQuantification.R
\name{SUPPZB}
\alias{SUPPZB}
\title{Cellular Quantification Domain Supplemental Variables}
\description{
{
  \tabular{ll}{
    \strong{Variable Name} \tab \strong{Variable Label} \cr
    STUDYID \tab Study Identifier \cr
    RDOMAIN  \tab Related Domain Abbreviation \cr
    USUBJID \tab Unique Subject Identifier \cr
    IDVAR \tab Identifying Variable \cr
    IDVARVAL \tab Identifying Variable Value \cr
    QNAM \tab Qualifier Variable Name \cr
    QLABEL \tab Qualifier Variable Label \cr
    QVAL \tab Data Value
  }
}
}
\note{
The following table enumerates the values in QNAM and QLABEL variables {
  \tabular{ll}{
    \strong{QNAM} \tab \strong{QLABEL} \cr
    ZBSPECSB \tab Specimen Subtype \cr
    ZBREFIDP \tab Source Specimen Identifier \cr
    ZBFCF \tab Control Files Names \cr
    VISITMIN \tab Planned Visit Minimum Start Day \cr
    VISITMAX \tab Planned Visit Maximum Start Day \cr
    ZBSPTRT \tab Specimen Treatment \cr
    ZBTRTAMV \tab Specimen Treatment Amount Value \cr
    ZBTRTAMU \tab Specimen Treatment Amount Unit \cr
    ZBTRTDUV \tab Specimen Treatment Duration Value \cr
    ZBTRTDUU \tab Specimen Treatment Duration Unit \cr
    ZBTRTTMV \tab Specimen Treatment Temperature Value \cr
    ZBTRTTMU \tab Specimen Treatment Temperature Unit
  }
}
}

