% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SubjectVisits.R
\name{SV}
\alias{SV}
\title{Subject Visits Domain Variables}
\description{
{
  \tabular{ll}{
    \strong{Variable Name  } \tab \strong{Variable Label} \cr
    STUDYID \tab Study Identifier \cr
    DOMAIN  \tab Domain Abbreviation \cr
    USUBJID \tab Unique Subject Identifier \cr
    VISITNUM \tab Visit Number \cr
    VISIT \tab Visit Name \cr
    SVSTDY \tab Study Day of Start of Visit
  }
}
}

