% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TrialInclusionExclusionCriteria.R
\name{TI}
\alias{TI}
\title{Trial Inclusion Exclusion Criteria Domain Variables}
\description{
{
  \tabular{ll}{
    \strong{Variable Name  } \tab \strong{Variable Label} \cr
    STUDYID \tab Study Identifier \cr
    DOMAIN  \tab Domain Abbreviation \cr
    IETEST \tab Inclusion/Exclusion Criterion \cr
    IECAT \tab Inclusion/Exclusion Category
  }
}
}

