% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TiterAssayResults.R
\name{Titer Assay Results Domain}
\alias{Titer Assay Results Domain}
\title{Titer Assay Results Domain}
\description{
The ImmPort study data generated from assays of types: HAI and Neut Ab Titer are grouped into the 
Titer Assay Results Domain. The data is reformated to a custom Titer Assay Results domain model in 
CDISC SDTM standards, and is a list of 2 data frames containing 1) Titer Assay Results data \code{\link{ZD}} 
and 2) any supplemental Titer Assay Results data \code{\link{SUPP}}
}

