% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TrialArms.R
\name{Trial Arms Domain}
\alias{Trial Arms Domain}
\title{Trial Arms Domain}
\description{
The Trial Arms data of an ImmPort study is reformated to the CDISC SDTM Trial Arms (TA) 
domain model, and is a list of 2 data frames containing 1) Trial Arms data \code{\link{TA}} 
and 2) any supplemental Trial Arms data \code{\link{SUPP}}
}

