% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TrialInclusionExclusionCriteria.R
\name{Trial Inclusion Exclusion Criteria Domain}
\alias{Trial Inclusion Exclusion Criteria Domain}
\title{Trial Inclusion Exclusion Criteria Domain}
\description{
The Trial Inclusion Exclusion Criteria data of an ImmPort study is reformated to the 
CDISC SDTM Trial Inclusion Exclusion Criteria (TI) domain model, and is a list of 2 data frames 
containing 1) Trial Inclusion Exclusion Criteria data \code{\link{TI}} 
and 2) any supplemental Trial Inclusion Exclusion Criteria data \code{\link{SUPP}}
}

