% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TrialSummary.R
\name{Trial Summary Domain}
\alias{Trial Summary Domain}
\title{Trial Summary Domain}
\description{
The Trial Summary data of an ImmPort study is reformated to the CDISC SDTM Trial Summary (TS) 
domain model, and is a list of 2 data frames containing 1) Trial Summary data \code{\link{TS}} 
and 2) any supplemental Trial Summary data \code{\link{SUPP}}
}

