% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TrialVisits.R
\name{Trial Visits Domain}
\alias{Trial Visits Domain}
\title{Trial Visits Domain}
\description{
Information on the planned visits of an ImmPort study is reformated to the CDISC SDTM Trial Visits (TV) 
domain model, and is a list of 2 data frames containing 1) Trial Visits data \code{\link{TV}} 
and 2) any supplemental Trial Visits data \code{\link{SUPP}}
}

