% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Study.R
\docType{class}
\name{TrialDesign-class}
\alias{TrialDesign}
\alias{TrialDesign-class}
\title{Trial Design class}
\description{
Trial Design class
}
\section{Fields}{

\describe{
\item{\code{ta_l}}{Trial Arms data \code{\link{TA}} and supplemental Trial Arms data \code{\link{SUPP}}}

\item{\code{ti_l}}{Trial Inclusion Exclusion Criteria data \code{\link{TI}} and supplemental Trial Inclusion Exclusion Criteria data \code{\link{SUPP}}}

\item{\code{ts_l}}{Trial Summary data \code{\link{TS}} and supplemental Trial Summary data \code{\link{SUPP}}}
}}
\examples{
library(DBI)
library(sqldf)
studies_dir <- system.file("extdata", "ImmPortStudies", package = "RImmPort")
db_dir <- file.path(studies_dir, "Db")
sqlite_conn <- dbConnect(SQLite(), dbname=file.path(db_dir, "ImmPort.sqlite"))
setImmPortDataSource(sqlite_conn)
sdy139 <- getStudy("SDY139")
ts_df <- sdy139$trial_design$ts_l$ts_df
}

