% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TrialSummary.R
\name{TS}
\alias{TS}
\title{Trial Summary Domain Variables}
\description{
{
  \tabular{ll}{
    \strong{Variable Name} \tab \strong{Variable Label} \cr
    STUDYID \tab Study Identifier \cr
    DOMAIN  \tab Domain Abbreviation \cr
    TSSEQ \tab Sequence Number \cr
    TSPARMCD \tab Trial Summary Parameter Short Name \cr
    TSPARM \tab Trial Summary Parameter \cr
    TSVAL \tab Parameter Value
  }
}
}
\note{
The following table enumerates the values in TSPARMCD and TSPARM variables {
  \tabular{ll}{
    \strong{TSPARMCD} \tab \strong{TSPARM} \cr
    TITLE \tab Trial Title \cr
    DESCR \tab Trial Description \cr
    INDIC \tab Trial Indication \cr
    TRT \tab Investigational Therapy or Treatment \cr
    HYPOTHS \tab Trial Hypotheses \cr
    SSTDTC \tab Study Start Date \cr
    SENDTC \tab Study End Date \cr
    PLANSUB \tab Planned Number of Subjects \cr
    ACTSUB \tab Actual Number of Subjects \cr
    AGEMAX \tab Planned Maximum Age of Subjects \cr
    AGEMIN \tab Planned Minimum Age of Subjects \cr
    AGEU \tab Age Units \cr
    SEXPOP \tab Sex of Participants \cr
    SPONSOR \tab Clinical Study Sponsor \cr
    PUBRLDAT \tab Public Release Date \cr
    ISTRIAL \tab Study Type \cr
    RESFOCUS \tab Trial Research Focus 
  }
}
}

