% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TrialVisits.R
\name{TV}
\alias{TV}
\title{Trial Visits Domain Variables}
\description{
{
  \tabular{ll}{
    \strong{Variable Name  } \tab \strong{Variable Label} \cr
    STUDYID \tab Study Identifier \cr
    DOMAIN  \tab Domain Abbreviation \cr
    VISITNUM \tab Visit Number \cr
    VISIT \tab Visit Name \cr
    ARMCD \tab Planned Arm Code \cr
    ARM \tab Name of Planned Arm \cr
    TVSTRL \tab Visit Start Rule \cr
    TVENRL \tab Visit End Rule
  }
}
}

