% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VitalSigns.R
\name{Vital Signs Domain}
\alias{Vital Signs Domain}
\title{Vital Signs Domain}
\description{
The Vital Signs data of an ImmPort study is reformated to the CDISC SDTM Vital Signs (VS) 
domain model, and is a list of 2 data frames containing 1) Vital Signs data \code{\link{VS}} 
and 2) any supplemental Vital Signs data \code{\link{SUPPVS}}
}

