% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/VitalSigns.R
\name{VS}
\alias{VS}
\title{Vital Signs Domain Variables}
\description{
{
  \tabular{ll}{
    \strong{Variable Name  } \tab \strong{Variable Label} \cr
    STUDYID \tab Study Identifier \cr
    DOMAIN  \tab Domain Abbreviation \cr
    USUBJID \tab Unique Subject Identifier \cr
    VSSEQ \tab Sequence Number \cr
    VSTEST \tab Vital Signs Test Name \cr
    VSCAT \tab Category for Vital Signs \cr
    VSORRES \tab Result or Finding in Original Units \cr
    VSORRESU \tab Original Units \cr
    VSLOC \tab Location of Vital Signs Measurement \cr
    VSDY \tab Study Day of Vital Signs
  }
}
}

