% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ProteinQuantification.R
\name{ZA}
\alias{ZA}
\title{Protein Quantification Domain Variables}
\description{
{
  \tabular{ll}{
    \strong{Variable Name  } \tab \strong{Variable Label} \cr
    STUDYID \tab Study Identifier \cr
    DOMAIN  \tab Domain Abbreviation \cr
    USUBJID \tab Unique Subject Identifier \cr
    ZASEQ \tab Sequence Number \cr
    ZATEST \tab Protein Quantification Test Name \cr
    ZACAT \tab Category for Protein Quantification \cr
    ZAMETHOD \tab Measurement Technique \cr
    ZAANALYT \tab Analyte \cr
    ZAORRES \tab Result or Finding in Original Units \cr
    ZAORRESU \tab Original Units \cr
    ZASPEC \tab Specimen Type \cr
    ZASPECSB \tab Specimen Subtype \cr
    ZADY \tab Study Time of Specimen Collection \cr
    ZADTCU \tab Units of Study Time of Specimen Collection \cr
    ZASPECID \tab Specimen Identifier \cr
    ZAXFN \tab Raw Data File or Life Science Identifier
  }
}
}

