% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/CellularQuantification.R
\name{ZB}
\alias{ZB}
\title{Cellular Quantification Domain Variables}
\description{
{
  \tabular{ll}{
    \strong{Variable Name  } \tab \strong{Variable Label} \cr
    STUDYID \tab Study Identifier \cr
    DOMAIN  \tab Domain Abbreviation \cr
    USUBJID \tab Unique Subject Identifier \cr
    ZBSEQ \tab Sequence Number \cr
    ZBTEST \tab Cellular Quantification Test Name \cr
    ZBCAT \tab Category for Cellular Quantification \cr
    ZBMETHOD \tab Measurement Technique \cr
    ZBPOPDEF \tab Cell Population Definition \cr
    ZBPOPNAM \tab Cell Population Name \cr
    ZBORRES \tab Result or Finding in Original Units \cr
    ZBORRESU \tab Original Units \cr
    ZBBASPOP \tab Base Parent Population \cr
    ZBSPEC \tab Specimen Type \cr
    ZBSPECSB \tab Specimen Subtype \cr
    ZBDY \tab Study Day of Specimen Collection \cr
    ZBDYU \tab Units of Study Day of Specimen Collection \cr
    ZBREFID \tab Specimen Identifier \cr
    ZBXFN \tab Raw Data File or Life Science Identifier
  }
}
}

