% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Study.R
\docType{class}
\name{Events-class}
\alias{Events}
\alias{Events-class}
\title{Events class}
\description{
Events class
}
\section{Fields}{

\describe{
\item{\code{ae_l}}{Adverse Events data \code{\link{AE}} and supplemental Adverse Events data \code{\link{SUPP}}}

\item{\code{dv_l}}{Protocol Deviations data \code{\link{DV}} and supplemental Protocol Deviations data \code{\link{SUPPDV}}}

\item{\code{mh_l}}{Medical History data \code{\link{MH}} and supplemental Medical History data \code{\link{SUPPMH}}}

\item{\code{apmh_l}}{Associated Persons Medical History data \code{\link{APMH}} and supplemental Associated Persons Medical History data \code{\link{SUPP}}}
}}
\examples{
library(DBI)
library(sqldf)
studies_dir <- system.file("extdata", "ImmPortStudies", package = "RImmPort")
db_dir <- file.path(studies_dir, "Db")
sqlite_conn <- dbConnect(SQLite(), dbname=file.path(db_dir, "ImmPort.sqlite"))
setImmPortDataSource(sqlite_conn)
sdy139 <- getStudy("SDY139")
ae_df <- sdy139$events$ae_l$ae_df
}

