% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Study.R
\name{getListOfStudies}
\alias{getListOfStudies}
\title{Get a list of all studies}
\usage{
getListOfStudies()
}
\value{
List of study indentifiers
}
\description{
Get a list of all studies
}
\examples{
library(DBI)
library(sqldf)
studies_dir <- system.file("extdata", "ImmPortStudies", package = "RImmPort")
db_dir <- file.path(studies_dir, "Db")
sqlite_conn <- dbConnect(SQLite(), dbname=file.path(db_dir, "ImmPort.sqlite"))
setImmPortDataSource(sqlite_conn)
study_ids <- getListOfStudies()
}

