% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PhysicalExamination.R
\name{PE}
\alias{PE}
\title{Physical Examination Domain Variables}
\description{
{
  \tabular{ll}{
    \strong{Variable Name  } \tab \strong{Variable Label} \cr
    STUDYID \tab Study Identifier \cr
    DOMAIN  \tab Domain Abbreviation \cr
    USUBJID \tab Unique Subject Identifier \cr
    PESEQ \tab Sequence Number \cr
    PETEST \tab Body System Examined \cr
    PECAT \tab Category for Examination \cr
    PEBODSYS \tab Body System or Organ Class \cr
    PEORRES \tab Verbatim Examination Finding \cr
    PEORRESU \tab Original Units \cr
    PELOC \tab Location of Physical Exam Finding \cr
    VISITNUM \tab Visit Number \cr
    VISIT \tab Visit Name \cr
    PEDY \tab Study Day of Examination
  }
}
}

