% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Study.R
\docType{class}
\name{SpecialPurpose-class}
\alias{SpecialPurpose}
\alias{SpecialPurpose-class}
\title{Special Purpose class}
\description{
Special Purpose class
}
\section{Fields}{

\describe{
\item{\code{dm_l}}{Demographics data \code{\link{DM}} and supplemental Demographics data \code{\link{SUPP}}}

\item{\code{sv_l}}{Subject Visits data \code{\link{SV}} and supplemental Subject Visits data \code{\link{SUPP}}}
}}
\examples{
library(DBI)
library(sqldf)
studies_dir <- system.file("extdata", "ImmPortStudies", package = "RImmPort")
db_dir <- file.path(studies_dir, "Db")
sqlite_conn <- dbConnect(SQLite(), dbname=file.path(db_dir, "ImmPort.sqlite"))
setImmPortDataSource(sqlite_conn)
sdy139 <- getStudy("SDY139")
dm_df <- sdy139$special_purpose$dm_l$dm_df
}

