% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SkinResponse.R
\name{SR}
\alias{SR}
\title{Skin Response Domain Variables}
\description{
{
  \tabular{ll}{
    \strong{Variable Name  } \tab \strong{Variable Label} \cr
    STUDYID \tab Study Identifier \cr
    DOMAIN  \tab Domain Abbreviation \cr
    USUBJID \tab Unique Subject Identifier \cr
    SRSEQ \tab Sequence Number \cr
    SRTEST \tab Skin Response Test or Examination Name \cr
    SROBJ \tab Object of the Observation \cr
    SRCAT \tab Category for Test \cr
    SRORRES \tab Results or Findings in Original Units \cr
    SRORRESU \tab Original Units \cr
    SRLOC \tab Location used for Measurement \cr
    VISITNUM \tab Visit Number \cr
    VISIT \tab Visit Name \cr
    SRDY \tab Study Day of Visit/Collection/Exam
  }
}
}

