% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TrialArms.R
\name{TA}
\alias{TA}
\title{Trial Arms Domain Variables}
\description{
{
  \tabular{ll}{
    \strong{Variable Name  } \tab \strong{Variable Label} \cr
    STUDYID \tab Study Identifier \cr
    DOMAIN  \tab Domain Abbreviation \cr
    ARMCD \tab Planned Arm Code \cr
    ARM \tab Name of Planned Arm \cr
    ARMDESC \tab Description of Planned Arm \cr
    ARMRULE \tab Population Selection Rule
  }
}
}

