% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ProteinQuantification.R
\name{ZA}
\alias{ZA}
\title{Protein Quantification Domain Variables}
\description{
{
  \tabular{ll}{
    \strong{Variable Name  } \tab \strong{Variable Label} \cr
    STUDYID \tab Study Identifier \cr
    DOMAIN  \tab Domain Abbreviation \cr
    USUBJID \tab Unique Subject Identifier \cr
    ZASEQ \tab Sequence Number \cr
    ZATEST \tab Protein Quantification Test Name \cr
    ZACAT \tab Category for Protein Quantification \cr
    ZAMETHOD \tab Measurement Technique \cr
    ZAANALYT \tab Analyte \cr
    ZAORRES \tab Result or Finding in Original Units \cr
    ZAORRESU \tab Original Units \cr
    ZASPEC \tab Specimen Type \cr
    VISITNUM \tab Visit Number \cr
    VISIT \tab Visit Name \cr
    ZAELTM \tab Planned Elapsed Time from Time Point Ref \cr
    ZATPTREF \tab Time Point Reference \cr
    ZAREFID \tab Specimen Identifier \cr
    ZAXFN \tab Raw Data File or Life Science Identifier
  }
}
}

