% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NucleicAcidQuantification.R
\name{ZC}
\alias{ZC}
\title{Nucleic Acid Quantification Domain Variables}
\description{
{
  \tabular{ll}{
    \strong{Variable Name  } \tab \strong{Variable Label} \cr
    STUDYID \tab Study Identifier \cr
    DOMAIN  \tab Domain Abbreviation \cr
    USUBJID \tab Unique Subject Identifier \cr
    ZCSEQ \tab Sequence Number \cr
    ZCTEST \tab Nucleic Acid Quantification Test Name \cr
    ZCCAT \tab Category for Nucleic Acid Quantification \cr
    ZCMETHOD \tab Measurement Technique \cr
    ZCPOPDEF \tab Cell Population Definition \cr
    ZCPOPNAM \tab Cell Population Name \cr
    ZCORRES \tab Result or Finding in Original Units \cr
    ZCORRESU \tab Original Units \cr
    ZCBASPOP \tab Base Parent Population \cr
    ZCSPEC \tab Specimen Type \cr
    ZCREFID \tab Specimen Identifier \cr
    VISITNUM \tab Visit Number \cr
    VISIT \tab Visit Name \cr
    ZCELTM \tab Planned Elapsed Time from Time Point Ref \cr
    ZCTPTREF \tab Time Point Reference
  }
}
}

