% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TiterAssayResults.R
\name{ZD}
\alias{ZD}
\title{Titer Assay Results Domain Variables}
\description{
{
  \tabular{ll}{
    \strong{Variable Name  } \tab \strong{Variable Label} \cr
    STUDYID \tab Study Identifier \cr
    DOMAIN  \tab Domain Abbreviation \cr
    USUBJID \tab Unique Subject Identifier \cr
    ZDSEQ \tab Sequence Number \cr
    ZDTEST \tab Titer Assay Results Test Name \cr
    ZDCAT \tab Category for Titer Assay Results \cr
    ZDMETHOD \tab Measurement Technique \cr
    ZDPOPDEF \tab Cell Population Definition \cr
    ZDPOPNAM \tab Cell Population Name \cr
    ZDORRES \tab Result or Finding in Original Units \cr
    ZDORRESU \tab Original Units \cr
    ZDBASPOP \tab Base Parent Population \cr
    ZDSPEC \tab Specimen Type \cr
    ZDREFID \tab Specimen Identifier \cr
    VISITNUM \tab Visit Number \cr
    VISIT \tab Visit Name \cr
    ZDELTM \tab Planned Elapsed Time from Time Point Ref \cr
    ZDTPTREF \tab Time Point Reference \cr
    ZDXFN \tab Raw Data File or Life Science Identifier
  }
}
}

